+++ examples/json_writer/json_writer.go
@@ -11,21 +11,21 @@
 *                                                                            *
 *****************************************************************************/
 
-package main
+package json_writer
 
 import (
 	"github.com/rticommunity/rticonnextdds-connector-go"
-	"log"
 	"path"
 	"runtime"
 	"time"
+	"log"
 )
 
 type Shape struct {
-	Color     string `json:"color"`
-	X         int    `json:"x"`
-	Y         int    `json:"y"`
-	Shapesize int    `json:"shapesize"`
+        Color     string `json:"color"`
+        X         int    `json:"x"`
+        Y         int    `json:"y"`
+        Shapesize int    `json:"shapesize"`
 }
 
 func main() {
@@ -44,30 +44,22 @@ func main() {
 	// Delete the connector when this main function returns
 	defer connector.Delete()
 
-	// Get an input from the connector
-	input, err := connector.GetInput("MySubscriber::MySquareReader")
+	// Get an output from the connector
+	output, err := connector.GetOutput("MyPublisher::MySquareWriter")
 	if err != nil {
 		log.Panic(err)
 	}
 
-	// Get values from a received sample and print them
+	// Set values to the instance and write the instance
 	for i := 0; i < 500; i++ {
-		input.Take()
-		numOfSamples := input.Samples.GetLength()
-		for j := 0; j < numOfSamples; j++ {
-			if input.Infos.IsValid(j) {
-				var shape Shape
-				err := input.Samples.Get(j, &shape)
-				if err != nil {
-					log.Println(err)
-				}
-				log.Println("---Received Sample---")
-				log.Printf("color: %s\n", shape.Color)
-				log.Printf("x: %d\n", shape.X)
-				log.Printf("y: %d\n", shape.Y)
-				log.Printf("shapesize: %d\n", shape.Shapesize)
-			}
-		}
+		var shape Shape
+		shape.X = i
+		shape.Y = i*2
+		shape.Shapesize = 30
+		shape.Color = "BLUE"
+		output.Instance.Set(&shape)
+		output.Write()
+		log.Println("Writing...")
 		time.Sleep(time.Second * 1)
 	}
 }
