+++ examples/read_and_write.go/read_and_write.go
