+++ examples/reader_wait/reader_wait.go
