+++ examples/simple_reader/reader.go
