+++ examples/simple_writer/writer.go
